LPD-48507 Create AttachmentBase and adapt Attachment component
- Make the Attachment have the Fieldbase and the corresponding props
- Rename the interfaces and types
- Create the FileContainer component outside AttachmentBase
- Create an util file with the validate functions
- AttachmentBase has now the main code that will be reused when AttachmentLocalizedObjectField is created
